Move DataImporter filesystem responsibilities to new FileSystem class
