+++ src/DataImporter.ts
@@ -1,6 +1,6 @@
 import { ObjectId } from 'mongodb';
-import { readdirSync, lstatSync } from 'fs';
 import { Database } from './Database';
+import { fileSystem } from './FileSystem';
 import { AppConfig } from './config';
 import { log } from './logger';
 
@@ -13,7 +13,6 @@ export interface CollectionToImport {
 
 export class DataImporter {
   static DIRECTORY_NAME_PATTERN_SEPARATOR = '-';
-  static FILE_NAME_SPLIT_CHARACTER = '.';
 
   constructor(public db: Database) {}
 
@@ -43,7 +42,7 @@ export class DataImporter {
     inputDirectory: string,
     existingCollections: string[] = [],
   ): CollectionToImport[] {
-    const collectionsDirectories = this.getDirectories(inputDirectory);
+    const collectionsDirectories = fileSystem.getDirectories(inputDirectory);
     const collectionsToImport = collectionsDirectories.map(
       collectionDirectory => {
         return this.getCollectionToImport(
@@ -62,15 +61,15 @@ export class DataImporter {
     collectionPath: string,
     config: AppConfig,
   ) {
-    const fileNames = readdirSync(collectionPath) || [];
-    const documentFileNames = this.getSupportedDocumentFileNames(
+    const fileNames = fileSystem.listFiles(collectionPath);
+    const documentFileNames = fileSystem.getSupportedDocumentFileNames(
       fileNames,
       config.supportedExtensions,
     );
 
     log(`Inserting documents into collection ${collectionName}...`);
 
-    const documentsContentArray = this.getDocumentsContentArray(
+    const documentsContentArray = fileSystem.getFilesContentArray(
       collectionPath,
       documentFileNames,
     );
@@ -104,62 +103,6 @@ export class DataImporter {
     existingCollections: string[] = [],
   ) {
     return existingCollections.indexOf(collection) === -1;
-  }
-
-  getDocumentsContentArray(
-    collectionPath: string,
-    documentFileNames: string[],
-  ) {
-    const documentsContentArray = documentFileNames.reduce<any[]>(
-      (arr: any[], documentFileName) => {
-        const documentPath = `${collectionPath}/${documentFileName}`;
-        const document: any = require(documentPath);
-        return arr.concat(document);
-      },
-      [],
-    );
-
-    return documentsContentArray;
-  }
-
-  getSupportedDocumentFileNames(
-    fileNames: string[],
-    supportedExtensions: string[],
-  ) {
-    const documentFileNames = fileNames.filter(fileName => {
-      const fileNameSplit = fileName.split(
-        DataImporter.FILE_NAME_SPLIT_CHARACTER,
-      );
-      if (this.shouldIgnoreFile(fileNameSplit)) {
-        return false;
-      }
-
-      const fileExtension = fileNameSplit.pop()!;
-      return supportedExtensions.some(
-        extension => extension.toLowerCase() === fileExtension.toLowerCase(),
-      );
-    });
-
-    return documentFileNames;
-  }
-
-  shouldIgnoreFile(fileNameSplitByDot: string[]) {
-    const isHidden = !fileNameSplitByDot[0];
-    const hasNoExtension = fileNameSplitByDot.length === 1;
-    return isHidden || hasNoExtension;
-  }
-
-  getDirectories(inputDirectory: string) {
-    const filesAndDirectories = readdirSync(inputDirectory);
-    const directories = filesAndDirectories.filter(fileOrDirectory =>
-      this.isDirectory(`${inputDirectory}/${fileOrDirectory}`),
-    );
-    return directories;
-  }
-
-  isDirectory(collectionPath: string) {
-    const stats = lstatSync(collectionPath);
-    return stats.isDirectory();
   }
 
   getCollectionName(directoryName: string) {
