+++ src/DatabaseConnector.ts
@@ -37,3 +37,5 @@ export class DatabaseConnector {
     return `${protocol}://${host}:${port}/${name}`;
   };
 }
+
+export const databaseConnector = new DatabaseConnector(MongoClient);
