+++ tests/DataImporter.ts
@@ -6,7 +6,7 @@ const dataImporter = new DataImporter(
   new Database(jest.genMockFromModule('mongodb')),
 );
 
-describe('Reading files', () => {
+describe('Processing data', () => {
   it('should get proper collection name from directory name', () => {
     const testCollectionName = 'test collection';
 
@@ -35,65 +35,6 @@ describe('Reading files', () => {
     expect(collectionWithNumberName).toBe('1');
   });
 
-  it('should filter supported documents from all files in directory', () => {
-    const fileNames = [
-      '.README.md',
-      'dist',
-      'file1.json',
-      'file2.js',
-      'release.sh',
-      'data-import.js',
-    ];
-    const supportedExtensions = ['js', 'json'];
-    const supportedDocuments = dataImporter.getSupportedDocumentFileNames(
-      fileNames,
-      supportedExtensions,
-    );
-    expect(supportedDocuments).toEqual([
-      'file1.json',
-      'file2.js',
-      'data-import.js',
-    ]);
-  });
-
-  it('should recognize hidden or files with no extension', () => {
-    const ignoreFile = dataImporter.shouldIgnoreFile(
-      'test.extension'.split('.'),
-    );
-    expect(ignoreFile).toBeFalsy();
-
-    const ignoreHiddenFile = dataImporter.shouldIgnoreFile(
-      '.test.js'.split('.'),
-    );
-    expect(ignoreHiddenFile).toBeTruthy();
-
-    const ignoreFileWithNoExtension = dataImporter.shouldIgnoreFile(
-      'test'.split('.'),
-    );
-    expect(ignoreFileWithNoExtension).toBeTruthy();
-
-    const ignoreHiddenFileWithNoExtension = dataImporter.shouldIgnoreFile(
-      '.test'.split('.'),
-    );
-    expect(ignoreHiddenFileWithNoExtension).toBeTruthy();
-  });
-});
-
-jest.mock('mockFiles/test1.json', () => ({ number: 1 }), { virtual: true });
-jest.mock(
-  'mockFiles/test2.js',
-  () => ({
-    value: {
-      second: true,
-    },
-  }),
-  { virtual: true },
-);
-jest.mock('mockFiles/test3.json', () => ({ string: 'three' }), {
-  virtual: true,
-});
-
-describe('Processing data', () => {
   it('should create collection only it does not exist', () => {
     const collection = 'testing';
     const existingCollections = ['just', 'a', 'simple', 'test'];
@@ -108,28 +49,6 @@ describe('Processing data', () => {
       existingCollections,
     );
     expect(shouldCreateTestCollection).toBeFalsy();
-  });
-
-  it('should get documents content array from array of file names', () => {
-    const documentFileNames = ['test1.json', 'test2.js', 'test3.json'];
-    const contentArray = dataImporter.getDocumentsContentArray(
-      'mockFiles',
-      documentFileNames,
-    );
-
-    expect(contentArray).toEqual([
-      {
-        number: 1,
-      },
-      {
-        value: {
-          second: true,
-        },
-      },
-      {
-        string: 'three',
-      },
-    ]);
   });
 
   it('should replace document `id` field with `_id`', () => {
